Components: Extract a reusable PostLastRevision component
