+++ editor/post-last-revision/check.js
@@ -1,0 +1,25 @@
+/**
+ * External dependencies
+ */
+import { connect } from 'react-redux';
+
+/**
+ * Internal dependencies
+ */
+import { getCurrentPostLastRevisionId } from '../selectors';
+
+function PostLastRevisionCheck( { lastRevisionId, children } ) {
+	if ( ! lastRevisionId ) {
+		return null;
+	}
+
+	return children;
+}
+
+export default connect(
+	( state ) => {
+		return {
+			lastRevisionId: getCurrentPostLastRevisionId( state ),
+		};
+	}
+)( PostLastRevisionCheck );
