+++ editor/post-last-revision/style.scss
@@ -1,4 +1,4 @@
-.editor-last-revision__title {
+.editor-post-last-revision__title {
 	padding: 0;
 	width: 100%;
 	font-weight: 600;
